Multiple Issues Fixed:
Below issues are fixed:

On a Final ROC email, no Fuel information needs to be provided.  Currently, the Final email displays fuel types from the previous ROC email.  Should not show any fuel types at all in the email.

The disclaimer at the top of the email: "Intel - For internal….." should not be bold

Bullet formating is incorrect - The bullets should not be indented.  Removing the indent for bullets will fix this issue.